Don't need to convert to int from Integer for AppdefEntityID

+++ src/org/hyperic/hq/ui/action/resource/common/inventory/ChangeResourceOwnerAction.java
@@ -6,7 +6,7 @@
  * normal use of the program, and does *not* fall under the heading of
  * "derived work".
  * 
- * Copyright (C) [2004, 2005, 2006], Hyperic, Inc.
+ * Copyright (C) [2004-2008], Hyperic, Inc.
  * This file is part of HQ.
  * 
  * HQ is free software; you can redistribute it and/or modify
@@ -25,29 +25,23 @@
 
 package org.hyperic.hq.ui.action.resource.common.inventory;
 
-import java.io.IOException;
 import java.util.HashMap;
 
 import javax.servlet.ServletContext;
-import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-
-import org.hyperic.hq.appdef.shared.AppdefEntityID;
-import org.hyperic.hq.authz.shared.PermissionException;
-import org.hyperic.hq.authz.shared.AuthzSubjectValue;
-import org.hyperic.hq.common.ApplicationException;
-import org.hyperic.hq.ui.Constants;
-import org.hyperic.hq.ui.action.BaseAction;
-import org.hyperic.hq.ui.util.ActionUtils;
-import org.hyperic.hq.ui.util.ContextUtils;
-import org.hyperic.hq.ui.util.RequestUtils;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.struts.action.ActionForm;
 import org.apache.struts.action.ActionForward;
 import org.apache.struts.action.ActionMapping;
+import org.hyperic.hq.appdef.shared.AppdefEntityID;
+import org.hyperic.hq.authz.shared.PermissionException;
+import org.hyperic.hq.ui.Constants;
+import org.hyperic.hq.ui.action.BaseAction;
+import org.hyperic.hq.ui.util.ContextUtils;
+import org.hyperic.hq.ui.util.RequestUtils;
 
 /**
  * An Action that changes the owner of a resource in the inventory.
@@ -87,8 +81,7 @@ public class ChangeResourceOwnerAction extends BaseAction {
         Integer sessionId = RequestUtils.getSessionId(request);
 
         AppdefEntityID entityId =
-            new AppdefEntityID(resourceType.intValue(),
-                               resourceId.intValue());
+            new AppdefEntityID(resourceType.intValue(), resourceId);
         log.trace("setting owner [" + ownerId + "] for resource [" +
                   entityId + "]");
         ContextUtils.getAppdefBoss(ctx)
